+++ sourcespec/ssp_inversion.py
@@ -31,6 +31,7 @@ from sourcespec.ssp_spectral_model import (spectral_model, objective_func,
 from sourcespec.ssp_util import mag_to_moment, select_trace, smooth
 from sourcespec.ssp_radiated_energy import radiated_energy
 from sourcespec.ssp_inversion_types import InitialValues, Bounds
+from sourcespec.ssp_grid_search import grid_search
 logger = logging.getLogger(__name__.split('.')[-1])
 
 
@@ -85,6 +86,18 @@ def _curve_fit(config, spec, weight, yerr, initial_values, bounds):
             accept_test=bounds
         )
         params_opt = res.x
+        # trick: use curve_fit() bounded to params_opt
+        # to get the covariance
+        _, params_cov = curve_fit(
+            spectral_model, freq_log, ydata,
+            p0=params_opt, sigma=yerr,
+            bounds=(params_opt-(1e-10), params_opt+(1e-10))
+        )
+    elif config.inv_algorithm == 'GS':
+        minimize_func = objective_func(freq_log, ydata, weight)
+        nsteps = (10, 200, 200)  # we do fewer steps in magnitude
+        params_opt = grid_search(
+            minimize_func, bounds.bounds, nsteps)
         # trick: use curve_fit() bounded to params_opt
         # to get the covariance
         _, params_cov = curve_fit(
@@ -251,7 +264,8 @@ def spectral_inversion(config, spec_st, weight_st):
     algorithm_messages = {
         'TNC': 'Using truncated Newton algorithm for inversion.',
         'LM': 'Using Levenberg-Marquardt algorithm for inversion.',
-        'BH': 'Using basin-hopping algorithm for inversion.'
+        'BH': 'Using basin-hopping algorithm for inversion.',
+        'GS': 'Using grid search for inversion.'
     }
     logger.info(algorithm_messages[config.inv_algorithm])